# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CMS",
        "description": "Todas las rutas de la API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/socios": {
            "get": {
                "tags": [
                    "Socios"
                ],
                "summary": "Obtener una lista de socios",
                "operationId": "a5e3840d2cf49a338c021b9d1c248e47",
                "responses": {
                    "200": {
                        "description": "Lista de socios obtenida con éxito"
                    },
                    "404": {
                        "description": "No se han encontrado socios"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Socios"
                ],
                "summary": "Registrar un nuevo socio",
                "operationId": "db1c0f7dd3e8aadb8c1175e39a4ccddb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "dni"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678P"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Socio registrado con éxito"
                    },
                    "403": {
                        "description": "No tiene permisos para realizar esta acción"
                    },
                    "500": {
                        "description": "Socio existente, pruebe otro dni"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/socios/{id}": {
            "get": {
                "tags": [
                    "Socios"
                ],
                "summary": "Obtener detalles de un socio específico",
                "operationId": "f8e3c8494374a95c50bd6be09b8f366f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del socio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del socio obtenidos con éxito"
                    },
                    "404": {
                        "description": "Socio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Socios"
                ],
                "summary": "Actualizar los datos de un socio",
                "operationId": "4c9bb738587384e9923f968f8accd879",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del socio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "dni"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678P"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Socio actualizado con éxito"
                    },
                    "403": {
                        "description": "No tiene permisos para realizar esta acción"
                    },
                    "404": {
                        "description": "Socio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Socios"
                ],
                "summary": "Eliminar un socio",
                "operationId": "837325c73107f9b627533d8cea061bd7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del socio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Socio eliminado con éxito"
                    },
                    "403": {
                        "description": "No tiene permisos para realizar esta acción"
                    },
                    "404": {
                        "description": "Socio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usa un token de acceso para autenticar las peticiones",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}